% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Black_Scholes_call.R
\name{Black_Scholes_call}
\alias{Black_Scholes_call}
\title{Black-Scholes Call Pricing}
\usage{
Black_Scholes_call(s = 100, k = 100, v = 0.2, tt = 252/252, r = 0.05)
}
\arguments{
\item{s}{Starting Stock Price}

\item{k}{Strike Price}

\item{v}{Volatility Of Stock - Stock Volatility Return on Annual Basis}

\item{tt}{Days Until Maturity - Annualized}

\item{r}{Risk Free Rate - Annualized}
}
\value{
Price of Black-Scholes European Call Option

Evaluate price of a 90 strike call option with 20 days to expiration
(one month trading) with a risk free rate of 5%, 20% volatility and
stock trading at 100 a share
}
\description{
This function provides pricing for a Black-Scholes Call Option
}
\examples{
Black_Scholes_call(100, 90, 0.05, 20/252, 0.20)
}
