% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECO.R
\name{ECO}
\alias{ECO}
\title{European Call Option Pricing}
\usage{
ECO(s = 100, k = 100, v = 0.2, tt = 252/252, r = 0.05, m = 20, numsim = 10^4)
}
\arguments{
\item{s}{Starting Stock Price}

\item{k}{Strike Price}

\item{v}{Volatility Of Stock - Stock Volatility Return on Annual Basis}

\item{tt}{Days Until Maturity - Annualized}

\item{r}{Risk Free Rate - Annualized}

\item{m}{Number of Price Intervals to be Calculated}

\item{numsim}{Number of Price Simulations Run}
}
\value{
Estimated Price of European Call Option, Standard Error and One Standard Deviation Confidence Interval
}
\description{
This function provides a crude Monte Carlo Pricing for a European Call Option
Inspiration from: https://berkorbay.github.io/fe522/02_Monte_Carlo_Simulation.html#antithetic-variates
and package(derivmkts)
}
\examples{
ECO(100, 90, 0.05, 20/252, 0.20, 20, 10^4)
Evaluate price of a 90 strike call option with 20 days to expiration 
(one month trading) and 20 price intervals with a risk free rate of 
5\%, 20\% volatility and stock trading at 100 a share via 
10000 simulated prices  
}
