% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECO_CV.R
\name{ECO_CV}
\alias{ECO_CV}
\title{European Call Option Pricing with Control Variate (Final Stock Price)}
\usage{
ECO_CV(
  s = 100,
  k = 100,
  v = 0.2,
  tt = 252/252,
  r = 0.05,
  numsim = 10^4,
  with_naive = TRUE
)
}
\arguments{
\item{s}{Starting Stock Price}

\item{k}{Strike Price}

\item{v}{Volatility Of Stock - Stock Volatility Return on Annual Basis}

\item{tt}{Days Until Maturity - Annualized}

\item{r}{Risk Free Rate - Annualized}

\item{numsim}{Number of Price Simulations Run}

\item{with_naive}{If True, return crude Monte Carlo Estimate for Comparision}
}
\value{
Estimated Price of European Call Option, Standard Error and One Standard Deviation Confidence Interval
}
\description{
This function provides Monte Carlo Pricing for a European Call Option with the Final Stock Price as a Control Variate
Inspiration from: https://berkorbay.github.io/fe522/02_Monte_Carlo_Simulation.html#antithetic-variates
and package(derivmkts)
}
\examples{
ECO_CV(100, 90, 0.05, 20/252, 0.20, 20, 10^4, FALSE)
Evaluate price of a 90 strike call option with 20 days to expiration 
(one month trading) and 20 price intervals with a risk free rate of 
5\%, 20\% volatility and stock trading at 100 a share via 
10000 simulated prices and no print out of Crude Monte Carlo Simulation
}
