% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CF_GEO.R
\name{GEO_AO}
\alias{GEO_AO}
\title{Closed Formed Solution of Geometric Average Asian Option}
\usage{
GEO_AO(s = 100, k = 100, v = 0.2, tt = 252/252, r = 0.05, m = 20)
}
\arguments{
\item{s}{Starting Stock Price}

\item{k}{Strike Price}

\item{v}{Volatility Of Stock - Stock Volatility Return on Annual Basis}

\item{tt}{Days Until Maturity - Annualized}

\item{r}{Risk Free Rate - Annualized}

\item{m}{Number of Price Intervals to be Calculated}
}
\value{
Estimated Price of Asian Call Option, Standard Error and One Standard Deviation Confidence Interval
}
\description{
This function provides a Closed Form Solution to the Geometric Asian Option
Inspiration from: https://berkorbay.github.io/fe522/02_Monte_Carlo_Simulation.html#antithetic-variates
and package(derivmkts)
}
\examples{
GEO_AO(100, 90, 0.05, 20/252, 0.20, 20, FALSE)
Method Generates volatility and interest rates that are plugged back into the Black-Scholes Formula to produce a Geometric Average Price Estimate
}
